// src/types/tailwindTypes.ts
export type Colors =
	| 'yellow'
	| 'blue'
	| 'green'
	| 'red'
	| 'white'
	| 'black'
	| 'main'
	| 'accent'
	| 'west-side'
	| 'transparent'

export type HTagLevels = 1 | 2 | 3 | 4 | 5 | 6
export type Sizes = 'xs' | 'sm' | 'md' | 'lg' | 'xl' | '2xl' | '3xl' | '4xl' | '5xl' | '6xl' | '7xl' | '8xl'
export type FontWeights =
	| 'thin'
	| 'extralight'
	| 'light'
	| 'normal'
	| 'medium'
	| 'semibold'
	| 'bold'
	| 'extrabold'
	| 'black'